+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `replication-delete` command
 * Add `replication-pause` command
 * Add `replication-unpause` command
+* Add `--include-existing-files` to `replication-setup`
 
 ### Fixed
 * Fix `replication-setup` default priority setter