+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Fix `replication-setup` default priority setter
+
+### Infrastructure
 * Fix warnings in tests
 
 ## [3.5.0] - 2022-07-27